hotel-assistant-backend/requirements.txt
fastapi==0.115.6
uvicorn==0.27.0.post1
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-multipart==0.0.7
httpx==0.28.1
litellm
chromadb
PyMuPDF>=1.24.0
pymongo
